Fix wrong sender id

+++ app/src/main/java/ua/nure/notifications/database/Types.kt
@@ -1,4 +1,4 @@
 package ua.nure.notifications.database
 
 // Database entity id type
-typealias DatabaseId = Int
+typealias DatabaseId = Long
